+++ .gitignore
@@ -10,6 +10,7 @@ venv/
 *.egg-info/
 .ruff_cache/
 python/
+dist/
 
 # Pytest
 .coverage
