+++ src/nmea_gps_emulator/__init__.py
@@ -1,0 +1,3 @@
+"""NMEA GPS Emulator - A GPS receiver emulator that generates NMEA 0183 sentences."""
+
+__version__ = "1.0.0"
