+++ src/nmea_gps_emulator/constants.py
