+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for nmea_gps_emulator package."""
